requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
click==8.1.7
dnspython==2.4.2
python-dotenv==1.0.0
fastapi==0.103.1
h11==0.14.0
idna==3.4
joblib==1.3.2
numpy==1.26.0
pandas==2.0.3
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==4.5.0
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.3.0
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.23.2
